v2.6.0

+++ classes/test.php
@@ -6,9 +6,8 @@
 
 	$files = scandir(".");
 
-	$skip = array(".", "..", "ModuleUnitTester.php", "UnitTester.php", "test.php");
+	$skip = array(".", "..", "ModuleUnitTester.php", "UnitTester.php", "test.php", "testEmailManager.php", ".git");
 	echo "<h1>Unit Tests</h1>\n";
-	$color = "#b1d8ff";
 	$numFiles = 0;
 	foreach ($files as $file) {
 		if (!in_array($file, $skip)) {
@@ -19,41 +18,44 @@
 	foreach ($files as $file) {
 		if (!in_array($file, $skip)) {
 			require_once(dirname(__FILE__)."/".$file);
-			echo "<h2 style='background-color: $color'>Examining $file</h2>\n";
-			$myClass = preg_replace("/.php$/i", "", $file);
+			$myClass = preg_replace("/.php$/i", "", $file)."Test";
 
-			try {
-				$obj = new $myClass($token, $server, $pid);
-			} catch (Exception $e) {
-				echo $e->getMessage()."<br>";
-			}
-			$tester = new UnitTester();
-			$tester->analyze($obj);
-
-			$badResults = $tester->getFailures();
-			$numBadResults = 0;
-			foreach ($badResults as $test => $ary) {
-				$numBadResults += count($ary);
-			}
-			if ($numBadResults > 0) {
-				echo "<h4 style='background-color: #ff7c7c;'>$numBadResults Failures</h4>\n";
-				echo "<div id='$myClass"."_results'>";
-			} else {
-				echo "<h4 style='background-color: #bdffb6;' onclick='$(\"#$myClass"."_results\").show();'>All Passed</h4>\n";
-				echo "<div id='$myClass"."_results' style='display: none;'>";
-			}
-
-			$results = $tester->getResults();
-			foreach ($results as $test => $ary) {
-				echo "<h3>$test</h3>\n";
-				foreach ($ary['results'] as $result) {
-					if (preg_match("/FALSE/i", $result)) {
-						echo "<span style='color: red;'>$result</span><br>\n";
-					} else if (preg_match("/TRUE/i", $result)) {
-						echo "$result<br>\n";
-					} 
+			if (class_exists($myClass)) {
+				echo "<h2 class='blue'>Examining $file</h2>\n";
+				try {
+					$obj = new $myClass($token, $server, $pid);
+				} catch (Exception $e) {
+					echo $e->getMessage()."<br>";
 				}
+				$tester = new UnitTester();
+				$tester->analyze($obj);
+
+				$badResults = $tester->getFailures();
+				$numBadResults = 0;
+				foreach ($badResults as $test => $ary) {
+					$numBadResults += count($ary);
+				}
+				if ($numBadResults > 0) {
+					echo "<h4 class='red'>$numBadResults Failures</h4>\n";
+					echo "<div id='$myClass"."_results'>";
+				} else {
+					echo "<h4 class='green' onclick='$(\"#$myClass"."_results\").show();'>All Passed</h4>\n";
+					echo "<div id='$myClass"."_results' style='display: none;'>";
+				}
+	
+				$results = $tester->getResults();
+				foreach ($results as $test => $ary) {
+					echo "<h3>$test</h3>\n";
+					foreach ($ary['results'] as $result) {
+						if (preg_match("/FALSE/i", $result)) {
+							echo "<span class='red'>ERROR $result</span><br>\n";
+						} else if (preg_match("/TRUE/i", $result)) {
+							echo "$result<br>\n";
+						} 
+					}
+				}
+				echo "</div>";
 			}
-			echo "</div>";
 		}
 	}
+	echo "<p>Done</p>\n";
